data/roop/requirements.txt
absl-py==2.0.0
addict==2.4.0
albumentations==1.3.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
bash_kernel==0.9.1
basicsr==1.4.2
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.4
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
cryptography==3.4.8
customtkinter==5.2.0
cycler==0.12.1
Cython==3.0.6
darkdetect==0.8.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
easydict==1.11
executing==2.0.1
facexlib==0.3.0
fastjsonschema==2.18.1
filelock==3.13.1
filterpy==1.4.5
flatbuffers==23.5.26
fonttools==4.46.0
fqdn==1.5.1
fsspec==2023.12.1
future==0.18.3
gast==0.4.0
gdown==4.7.1
gfpgan==1.3.8
google-auth==2.23.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.9.0
httplib2==0.20.2
humanfriendly==10.0
idna==3.4
imageio==2.33.1
importlib-metadata==7.0.0
iniconfig==2.0.0
insightface==0.7.3
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-archive==3.4.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
keras==2.13.1
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.41.1
lmdb==1.4.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==8.10.0
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nbzip==0.1.0
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.0
onnx==1.14.0
onnxruntime-gpu==1.15.1
opencv-python==4.8.0.74
opencv-python-headless==4.8.1.78
opennsfw2==0.10.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
Pillow==10.0.0
platformdirs==3.11.0
pluggy==1.3.0
prettytable==3.9.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==7.4.3
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
qudida==0.0.4
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tb-nightly==2.16.0a20231210
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.18.0
tf-keras-nightly==2.16.0.dev2023121010
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2==1.2.1
tk==0.1.0
tkinterdnd2==0.3.0
tomli==2.0.1
torch==2.1.1+cu118
torchvision==0.16.1+cu118
tornado==6.3.3
tqdm==4.65.0
traitlets==5.13.0
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.5
wadllib==1.3.6
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.3.7
widgetsnbextension==4.0.9
wrapt==1.14.1
yapf==0.40.2
zipp==1.0.0

